refactor: move Sound.kt out of libanki
The equivalent to `sound.py` at Anki Desktop is actually SoundKt.kt, so there is no reason to keep `Sound.kt` at libanki
