+++ AnkiDroid/src/main/java/com/ichi2/anki/cardviewer/TTS.kt
@@ -22,8 +22,8 @@ import android.content.Context
 import com.ichi2.anki.CardUtils
 import com.ichi2.anki.R
 import com.ichi2.anki.ReadText
+import com.ichi2.anki.Sound.SoundSide
 import com.ichi2.libanki.Card
-import com.ichi2.libanki.Sound.SoundSide
 import com.ichi2.libanki.TTSTag
 import com.ichi2.libanki.Utils
 import com.ichi2.libanki.template.TemplateFilters
